+++ watertap/unit_models/zero_order/CANDOP_zo.py
@@ -14,6 +14,7 @@
 This module contains a zero-order representation of a CANDO+P reactor unit.
 """
 
+import pyomo.environ as pyo
 from pyomo.environ import Var, units as pyunits
 from idaes.core import declare_process_block_class
 from watertap.core import build_sido_reactive, ZeroOrderBaseData
@@ -100,3 +101,54 @@ class CANDOPData(ZeroOrderBaseData):
             return b.O2_demand[t] == (
                 b.extent_of_reaction[t, "n_reaction"] * b.oxygen_nitrogen_ratio
             )
+
+    @property
+    def default_costing_method(self):
+        return self.cost_CANDOP
+
+    @staticmethod
+    def cost_CANDOP(blk):
+        """
+        General method for costing CANDO+P reactor.
+        """
+        t0 = blk.flowsheet().time.first()
+
+        # Get parameter dict from database
+        parameter_dict = blk.unit_model.config.database.get_unit_operation_parameters(
+            blk.unit_model._tech_type, subtype=blk.unit_model.config.process_subtype
+        )
+
+        # Get costing parameter sub-block for this technology
+        size_param, size_cost = blk.unit_model._get_tech_parameters(
+            blk,
+            parameter_dict,
+            blk.unit_model.config.process_subtype,
+            ["sizing_parameter", "sizing_cost"],
+        )
+
+        # Add cost variable and constraint
+        blk.capital_cost = pyo.Var(
+            initialize=1,
+            units=blk.config.flowsheet_costing_block.base_currency,
+            bounds=(0, None),
+            doc="Capital cost of unit operation",
+        )
+
+        expr = pyo.units.convert(
+            blk.unit_model.properties_in[t0].flow_vol * size_param * size_cost,
+            to_units=blk.config.flowsheet_costing_block.base_currency,
+        )
+
+        # Determine if a costing factor is required
+        blk.unit_model._add_cost_factor(
+            blk, parameter_dict["capital_cost"]["cost_factor"]
+        )
+
+        blk.capital_cost_constraint = pyo.Constraint(
+            expr=blk.capital_cost == blk.cost_factor * expr
+        )
+
+        # Register flows
+        blk.config.flowsheet_costing_block.cost_flow(
+            blk.unit_model.electricity[t0], "electricity"
+        )
