+++ watertap/unit_models/zero_order/chemical_addition_zo.py
@@ -91,3 +91,57 @@ class ChemicalAdditionZOData(ZeroOrderBaseData):
         )
 
         pump_electricity(self, self.chemical_flow_vol)
+
+    @property
+    def default_costing_method(self):
+        return self.cost_chemical_addition
+
+    @staticmethod
+    def cost_chemical_addition(blk, number_of_parallel_units=1):
+        """
+        General method for costing chemical addition processes. Capital cost is
+        based on the mass flow rate of chemical added.
+        This method also registers the chemical flow and electricity demand as
+        costed flows.
+        Args:
+            number_of_parallel_units (int, optional) - cost this unit as
+                        number_of_parallel_units parallel units (default: 1)
+        """
+        chem_name = blk.unit_model.config.process_subtype
+
+        t0 = blk.flowsheet().time.first()
+        chem_flow_mass = (
+            blk.unit_model.chemical_dosage[t0]
+            * blk.unit_model.properties[t0].flow_vol
+            / blk.unit_model.ratio_in_solution
+        )
+        sizing_term = blk.unit_model.chemical_flow_vol[t0] / (
+            pyo.units.gal / pyo.units.day
+        )
+
+        # Get parameter dict from database
+        parameter_dict = blk.unit_model.config.database.get_unit_operation_parameters(
+            blk.unit_model._tech_type, subtype=blk.unit_model.config.process_subtype
+        )
+
+        # Get costing parameter sub-block for this technology
+        A, B = blk.unit_model._get_tech_parameters(
+            blk,
+            parameter_dict,
+            blk.unit_model.config.process_subtype,
+            ["capital_a_parameter", "capital_b_parameter"],
+        )
+
+        # Determine if a costing factor is required
+        factor = parameter_dict["capital_cost"]["cost_factor"]
+
+        # Call general power law costing method
+        blk.unit_model._general_power_law_form(
+            blk, A, B, sizing_term, factor, number_of_parallel_units
+        )
+
+        # Register flows
+        blk.config.flowsheet_costing_block.cost_flow(
+            blk.unit_model.electricity[t0], "electricity"
+        )
+        blk.config.flowsheet_costing_block.cost_flow(chem_flow_mass, chem_name)
