+++ watertap/unit_models/zero_order/chlorination_zo.py
@@ -14,6 +14,7 @@
 This module contains a zero-order representation of a Chlorination unit.
 """
 
+import pyomo.environ as pyo
 from pyomo.environ import units as pyunits, Var
 from idaes.core import declare_process_block_class
 from watertap.core import build_siso, constant_intensity, ZeroOrderBaseData
@@ -86,3 +87,77 @@ class ChlorinationZOData(ZeroOrderBaseData):
         self._perf_var_dict[
             "Chlorine Decay Rate (mg/(L*hr))"
         ] = self.chlorine_decay_rate
+
+    @property
+    def default_costing_method(self):
+        return self.cost_chlorination
+
+    @staticmethod
+    def cost_chlorination(blk):
+        """
+        General method for costing chlorination units. Capital cost is based on
+        the both inlet flow and dosage of chlorine.
+        This method also registers the chemical flow and electricity demand as
+        costed flows.
+        """
+        t0 = blk.flowsheet().time.first()
+        chem_flow_mass = (
+            blk.unit_model.chlorine_dose[t0] * blk.unit_model.properties_in[t0].flow_vol
+        )
+
+        # Get parameter dict from database
+        parameter_dict = blk.unit_model.config.database.get_unit_operation_parameters(
+            blk.unit_model._tech_type, subtype=blk.unit_model.config.process_subtype
+        )
+
+        # Get costing parameter sub-block for this technology
+        A, B, C = blk.unit_model._get_tech_parameters(
+            blk,
+            parameter_dict,
+            blk.unit_model.config.process_subtype,
+            ["capital_a_parameter", "capital_b_parameter", "capital_c_parameter"],
+        )
+
+        # Determine if a costing factor is required
+        factor = parameter_dict["capital_cost"]["cost_factor"]
+
+        # Add cost variable and constraint
+        blk.capital_cost = pyo.Var(
+            initialize=1,
+            units=blk.config.flowsheet_costing_block.base_currency,
+            bounds=(0, None),
+            doc="Capital cost of unit operation",
+        )
+
+        ln_Q = pyo.log(
+            pyo.units.convert(
+                blk.unit_model.properties_in[t0].flow_vol
+                / (pyo.units.m**3 / pyo.units.hour),
+                to_units=pyo.units.dimensionless,
+            )
+        )
+        ln_D = pyo.log(
+            pyo.units.convert(
+                blk.unit_model.chlorine_dose[t0] / (pyo.units.mg / pyo.units.liter),
+                to_units=pyo.units.dimensionless,
+            )
+        )
+
+        expr = pyo.units.convert(
+            A * ln_Q + B * ln_D + C * ln_Q * ln_D,
+            to_units=blk.config.flowsheet_costing_block.base_currency,
+        )
+
+        blk.unit_model._add_cost_factor(
+            blk, parameter_dict["capital_cost"]["cost_factor"]
+        )
+
+        blk.capital_cost_constraint = pyo.Constraint(
+            expr=blk.capital_cost == blk.cost_factor * expr
+        )
+
+        # Register flows
+        blk.config.flowsheet_costing_block.cost_flow(
+            blk.unit_model.electricity[t0], "electricity"
+        )
+        blk.config.flowsheet_costing_block.cost_flow(chem_flow_mass, "chlorine")
