+++ watertap/unit_models/zero_order/deep_well_injection_zo.py
@@ -14,6 +14,7 @@
 This module contains a zero-order representation of a deep well injection unit.
 """
 
+import pyomo.environ as pyo
 from pyomo.environ import Reference, units as pyunits, Var
 from idaes.core import declare_process_block_class
 
@@ -58,3 +59,76 @@ class DeepWellInjectionZOData(ZeroOrderBaseData):
 
         self._perf_var_dict["Pipe Distance (miles)"] = self.pipe_distance
         self._perf_var_dict["Pipe Diameter (inches)"] = self.pipe_diameter
+
+    @property
+    def default_costing_method(self):
+        return self.cost_deep_well_injection
+
+    @staticmethod
+    def cost_deep_well_injection(blk, number_of_parallel_units=1):
+        """
+        General method for costing deep well injection processes. Capital cost
+        is based on the cost of pump and pipe.
+        This method also registers the electricity demand as a costed flow.
+        Args:
+            number_of_parallel_units (int, optional) - cost this unit as
+                        number_of_parallel_units parallel units (default: 1)
+        """
+        t0 = blk.flowsheet().time.first()
+
+        # Get parameter dict from database
+        parameter_dict = blk.unit_model.config.database.get_unit_operation_parameters(
+            blk.unit_model._tech_type, subtype=blk.unit_model.config.process_subtype
+        )
+
+        # Get costing parameter sub-block for this technology
+        A, B, C = blk.unit_model._get_tech_parameters(
+            blk,
+            parameter_dict,
+            blk.unit_model.config.process_subtype,
+            ["well_pump_cost", "pipe_cost_basis", "flow_exponent"],
+        )
+
+        # Add cost variable and constraint
+        blk.capital_cost = pyo.Var(
+            initialize=1,
+            units=blk.config.flowsheet_costing_block.base_currency,
+            bounds=(0, None),
+            doc="Capital cost of unit operation",
+        )
+
+        cost_well_pump = A
+
+        cost_pipe = (
+            B * blk.unit_model.pipe_distance[t0] * blk.unit_model.pipe_diameter[t0]
+        )
+
+        cost_total = pyo.units.convert(
+            cost_well_pump + cost_pipe,
+            to_units=blk.config.flowsheet_costing_block.base_currency,
+        )
+
+        Q = pyo.units.convert(
+            blk.unit_model.properties[t0].flow_vol,
+            to_units=pyo.units.m**3 / pyo.units.hour,
+        )
+
+        sizing_term = Q / blk.unit_model.flow_basis[t0]
+
+        # Determine if a costing factor is required
+        factor = parameter_dict["capital_cost"]["cost_factor"]
+
+        # Call general power law costing method
+        blk.unit_model._general_power_law_form(
+            blk,
+            cost_total,
+            C,
+            sizing_term,
+            factor,
+            number_of_parallel_units,
+        )
+
+        # Register flows
+        blk.config.flowsheet_costing_block.cost_flow(
+            blk.unit_model.electricity[t0], "electricity"
+        )
