+++ watertap/unit_models/zero_order/landfill_zo.py
@@ -63,3 +63,46 @@ class LandfillZOData(ZeroOrderBaseData):
 
         self._perf_var_dict["Capacity Basis (kg/hr)"] = self.capacity_basis
         self._perf_var_dict["Total Mass (kg/hr)"] = self.total_mass
+
+    @property
+    def default_costing_method(self):
+        return self.cost_landfill
+
+    @staticmethod
+    def cost_landfill(blk, number_of_parallel_units=1):
+        """
+        General method for costing landfill. Capital cost is based on the total mass and
+        capacity basis.
+        Args:
+            number_of_parallel_units (int, optional) - cost this unit as
+                        number_of_parallel_units parallel units (default: 1)
+        """
+
+        t0 = blk.flowsheet().time.first()
+        sizing_term = blk.unit_model.total_mass[t0] / blk.unit_model.capacity_basis[t0]
+
+        # Get parameter dict from database
+        parameter_dict = blk.unit_model.config.database.get_unit_operation_parameters(
+            blk.unit_model._tech_type, subtype=blk.unit_model.config.process_subtype
+        )
+
+        # Get costing parameter sub-block for this technology
+        A, B = blk.unit_model._get_tech_parameters(
+            blk,
+            parameter_dict,
+            blk.unit_model.config.process_subtype,
+            ["capital_a_parameter", "capital_b_parameter"],
+        )
+
+        # Determine if a costing factor is required
+        factor = parameter_dict["capital_cost"]["cost_factor"]
+
+        # Call general power law costing method
+        blk.unit_model._general_power_law_form(
+            blk, A, B, sizing_term, factor, number_of_parallel_units
+        )
+
+        # Register flows
+        blk.config.flowsheet_costing_block.cost_flow(
+            blk.unit_model.electricity[t0], "electricity"
+        )
