+++ watertap/unit_models/zero_order/membrane_evaporator_zo.py
@@ -15,6 +15,7 @@
 This module contains a zero-order representation of a membrane evaporation.
 """
 
+import pyomo.environ as pyo
 from pyomo.environ import Var, units as pyunits
 from idaes.core import declare_process_block_class
 from watertap.core import build_sido, constant_intensity, ZeroOrderBaseData
@@ -62,3 +63,56 @@ class MembraneEvaporatorData(ZeroOrderBaseData):
                     b.water_flux * b.membrane_area, to_units=pyunits.m**3 / pyunits.s
                 )
             )
+
+    @property
+    def default_costing_method(self):
+        return self.cost_membrane_evaporator
+
+    @staticmethod
+    def cost_membrane_evaporator(blk):
+        """
+        General method for costing membrane evaporator unit.
+        """
+        t0 = blk.flowsheet().time.first()
+
+        # Get parameter dict from database
+        parameter_dict = blk.unit_model.config.database.get_unit_operation_parameters(
+            blk.unit_model._tech_type, subtype=blk.unit_model.config.process_subtype
+        )
+
+        # Get costing parameter sub-block for this technology
+        memb_cost = blk.unit_model._get_tech_parameters(
+            blk,
+            parameter_dict,
+            blk.unit_model.config.process_subtype,
+            ["membrane_cost"],
+        )
+
+        # Add cost variable and constraint
+        blk.capital_cost = pyo.Var(
+            initialize=1,
+            units=blk.config.flowsheet_costing_block.base_currency,
+            bounds=(0, None),
+            doc="Capital cost of unit operation",
+        )
+
+        expr = pyo.units.convert(
+            blk.unit_model.membrane_area * memb_cost,
+            to_units=blk.config.flowsheet_costing_block.base_currency,
+        )
+
+        # Determine if a costing factor is required
+        blk.unit_model._add_cost_factor(
+            blk, parameter_dict["capital_cost"]["cost_factor"]
+        )
+
+        blk.capital_cost_constraint = pyo.Constraint(
+            expr=blk.capital_cost == blk.cost_factor * expr
+        )
+
+        # Register flows
+        # TODO: Consider adding heat as a registered flow, since the inlet
+        #       stream to the membrane evaporator must be heated to ~37 C.
+        blk.config.flowsheet_costing_block.cost_flow(
+            blk.unit_model.electricity[t0], "electricity"
+        )
