+++ watertap/unit_models/zero_order/sedimentation_zo.py
@@ -15,6 +15,7 @@ This module contains a zero-order representation of a sedimentation unit
 operation.
 """
 
+import pyomo.environ as pyo
 from pyomo.environ import Constraint, units as pyunits, Var
 from idaes.core import declare_process_block_class
 
@@ -144,3 +145,104 @@ class SedimentationZOData(ZeroOrderBaseData):
                     "One of the following should be specified in the solute_list: "
                     "tss or phosphates"
                 )
+
+    @property
+    def default_costing_method(self):
+        return self.cost_sedimentation
+
+    @staticmethod
+    def cost_sedimentation(blk, number_of_parallel_units=1):
+        """
+        General method for costing sedimentaion processes. Capital cost is
+        based on the surface area of the basin.
+        Args:
+            number_of_parallel_units (int, optional) - cost this unit as
+                        number_of_parallel_units parallel units (default: 1)
+        """
+        t0 = blk.flowsheet().time.first()
+
+        if blk.unit_model.config.process_subtype != "phosphorus_capture":
+            sizing_term = blk.unit_model.basin_surface_area[t0] / pyo.units.foot**2
+
+            # Get parameter dict from database
+            parameter_dict = (
+                blk.unit_model.config.database.get_unit_operation_parameters(
+                    blk.unit_model._tech_type,
+                    subtype=blk.unit_model.config.process_subtype,
+                )
+            )
+
+            A, B = blk.unit_model._get_tech_parameters(
+                blk,
+                parameter_dict,
+                blk.unit_model.config.process_subtype,
+                ["capital_a_parameter", "capital_b_parameter"],
+            )
+
+            # Determine if a costing factor is required
+            factor = parameter_dict["capital_cost"]["cost_factor"]
+
+            # Call general power law costing method
+            blk.unit_model._general_power_law_form(
+                blk, A, B, sizing_term, factor, number_of_parallel_units
+            )
+        else:
+            # Get parameter dict from database
+            parameter_dict = (
+                blk.unit_model.config.database.get_unit_operation_parameters(
+                    blk.unit_model._tech_type,
+                    subtype=blk.unit_model.config.process_subtype,
+                )
+            )
+
+            # Get costing parameter sub-block for this technology
+            unit_capex, unit_opex = blk.unit_model._get_tech_parameters(
+                blk,
+                parameter_dict,
+                blk.unit_model.config.process_subtype,
+                ["unit_capex", "unit_opex"],
+            )
+
+            # Add cost variable and constraint
+            blk.capital_cost = pyo.Var(
+                initialize=1,
+                units=blk.config.flowsheet_costing_block.base_currency,
+                bounds=(0, None),
+                doc="Capital cost of unit operation",
+            )
+
+            capex_expr = pyo.units.convert(
+                blk.unit_model.properties_in[t0].flow_vol * unit_capex,
+                to_units=blk.config.flowsheet_costing_block.base_currency,
+            )
+
+            # Determine if a costing factor is required
+            blk.unit_model._add_cost_factor(
+                blk, parameter_dict["capital_cost"]["cost_factor"]
+            )
+
+            blk.capital_cost_constraint = pyo.Constraint(
+                expr=blk.capital_cost == blk.cost_factor * capex_expr
+            )
+
+            # Add fixed operating cost variable and constraint
+            blk.fixed_operating_cost = pyo.Var(
+                initialize=1,
+                units=blk.config.flowsheet_costing_block.base_currency
+                / blk.config.flowsheet_costing_block.base_period,
+                bounds=(0, None),
+                doc="Fixed operating cost of unit",
+            )
+            blk.fixed_operating_cost_constraint = pyo.Constraint(
+                expr=blk.fixed_operating_cost
+                == pyo.units.convert(
+                    blk.unit_model.properties_in[t0].flow_vol * unit_opex,
+                    to_units=blk.config.flowsheet_costing_block.base_currency
+                    / blk.config.flowsheet_costing_block.base_period,
+                )
+            )
+
+        # Register flows
+        blk.config.flowsheet_costing_block.cost_flow(
+            blk.unit_model.electricity[t0], "electricity"
+        )
