+++ watertap/unit_models/zero_order/supercritical_salt_precipitation_zo.py
@@ -14,6 +14,7 @@
 This module contains a zero-order representation of a supercritical salt precipitation unit.
 """
 
+import pyomo.environ as pyo
 from pyomo.environ import units as pyunits, Var
 from idaes.core import declare_process_block_class
 from watertap.core import build_sido, ZeroOrderBaseData
@@ -84,3 +85,63 @@ class SaltPrecipitationZOData(ZeroOrderBaseData):
 
         self._fixed_perf_vars.append(self.energy_electric_flow_mass)
         self._perf_var_dict["Electricity Intensity"] = self.energy_electric_flow_mass
+
+    @property
+    def default_costing_method(self):
+        return self.cost_supercritical_salt_precipitation
+
+    @staticmethod
+    def cost_supercritical_salt_precipitation(blk):
+        """
+        General method for costing supercritical salt precipitation unit.
+        """
+        t0 = blk.flowsheet().time.first()
+
+        # Get parameter dict from database
+        parameter_dict = blk.unit_model.config.database.get_unit_operation_parameters(
+            blk.unit_model._tech_type, subtype=blk.unit_model.config.process_subtype
+        )
+
+        # Get costing parameter sub-block for this technology
+        A, B, C, D = blk.unit_model._get_tech_parameters(
+            blk,
+            parameter_dict,
+            blk.unit_model.config.process_subtype,
+            [
+                "installation_factor",
+                "equipment_cost",
+                "base_flowrate",
+                "scaling_exponent",
+            ],
+        )
+
+        sizing_term = pyo.units.convert(
+            (blk.unit_model.flow_mass_in[t0] / C),
+            to_units=pyo.units.dimensionless,
+        )
+
+        # Add cost variable and constraint
+        blk.capital_cost = pyo.Var(
+            initialize=1,
+            units=blk.config.flowsheet_costing_block.base_currency,
+            bounds=(0, None),
+            doc="Capital cost of unit operation",
+        )
+
+        expr = pyo.units.convert(
+            A * B * sizing_term**D,
+            to_units=blk.config.flowsheet_costing_block.base_currency,
+        )
+
+        blk.unit_model._add_cost_factor(
+            blk, parameter_dict["capital_cost"]["cost_factor"]
+        )
+
+        blk.capital_cost_constraint = pyo.Constraint(
+            expr=blk.capital_cost == blk.cost_factor * expr
+        )
+
+        # Register flows
+        blk.config.flowsheet_costing_block.cost_flow(
+            blk.unit_model.electricity[t0], "electricity"
+        )
